run_qemu: restrict rootfs bootloader changes to fedora only
The bootloader fixups done in 'update_rootfs_boot_kernel' are Fedora
spicific - these allow the image created to boot normally as well,
outside of a run_qemu environment. Don't attempt these if the distro is
something other than fedora for now. We'll need to fix this for other
distros as needed. Address #2.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>