run_qemu: fix nfit_test cmdline reservations
The memmap reservation starting at 5G showed a resource conflict during
module reload. Fix this by starting the memmap reservations at 6G, and
reducing the sizes to 1G, 4G for the two namespaces.

The efi_fake_mem reservation was plain bogus with a size that was larger
than the available memory in the nfit_test config. Fix this to 1G to
fix.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>