run_qemu: Add option to skip ndctl build
When testing non-ndctl code while at the same time having ndctl changes
in progress it is beneficial to temporarily skip the ndctl build/install
in the root image.  This is because failures of the ndctl build will
stop running qemu despite not needing ndctl for the current test
required.

Add an option to skip ndctl build/install within the root image.  To
maintain current compatibility enable it by default.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>