run_qemu: Fix topology command line settings
Users are able to change the topology settings of number of memory
devices, pmem devices, nvmems, etc in two different ways.  First by use of
a preset of tiny, small, med, etc.  And secondly via individual options.

The preset option was broken in a1ac763bef57 because it would always
override the preset values with the defaults of the individual cmd line
options.

Fix the preset topology settings by making the individual options
override the preset settings only if specified.  This means the
individual options no longer have a specified preset.  Include nvmes as
part of the set of topology options for consistency.

Fixes: a1ac763bef57 ("run_qemu: Initialize variables with presets then user provided")
Cc: Reinette Chatre <reinette.chatre@intel.com>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>