run_qemu.sh: do not look for unused OVMF binaries on --legacy-bios.
This is a correction of previous commit 8504b014d7e6 ("run_qemu.sh: move
get_ovmf_binaries() later, to start_qemu()")

This is not so much about --legacy-bios but rather to regroup all
OVMF-related stuff in a single place and makes it easier to temporarily
disable OVMF locally. For instance: when using --direct-kernel which is
the default. And yes: the image can boot with --direct-kernel and no
OVMF at all. Whether that's a good idea or not is a different question -
but it's sometimes useful as a local test hack.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>