run_qemu.sh: move if !kvm accel=tcg test where it's used
Zero functional change.

It makes sense to have the sometimes complex dependencies and overrides
between options happening all in one place, in the
process_options_logic() function. But not when an option is unaffected
by other options and is merely changing the value of a variable. In that
case, it's more readable to have the logic in the same place where it's
used.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>